# backend-gin/internal/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/places": {
            "post": {
                "security": [
                    {
                        "OAuth2Password": [
                            "admin"
                        ]
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Place"
                ],
                "summary": "Place creation",
                "parameters": [
                    {
                        "minLength": 1,
                        "type": "string",
                        "example": "Fulham road",
                        "description": "The place address",
                        "name": "address",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "example": 123456789,
                        "description": "The ID of the place creator",
                        "name": "creatorId",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "minLength": 10,
                        "type": "string",
                        "example": "Stadium of Chelsea football club",
                        "description": "The place description, 10 characters minimum",
                        "name": "description",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "number",
                        "example": 51.48180425016331,
                        "description": "The latitude of the place",
                        "name": "lat",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "number",
                        "example": -0.19090418688755467,
                        "description": "The longitude of the place",
                        "name": "lng",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "minLength": 10,
                        "type": "string",
                        "example": "Stamford Bridge",
                        "description": "The place title/name, 10 characters minimum",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Place Image (JPEG)",
                        "name": "image",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.PlaceRead"
                        }
                    }
                }
            }
        },
        "/api/places/": {
            "get": {
                "security": [
                    {
                        "OAuth2Password": [
                            "admin"
                        ]
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Place"
                ],
                "summary": "Search and Filter places",
                "parameters": [
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "example": [
                            "ilike:Boulevard"
                        ],
                        "description": "The place address",
                        "name": "address",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "example": [
                            "in:123456789"
                        ],
                        "description": "The ID of the place creator",
                        "name": "creatorId",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "example": [
                            "like:football"
                        ],
                        "description": "The place description, 10 characters minimum",
                        "name": "description",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "enum": [
                                "id",
                                "title",
                                "description",
                                "address",
                                "location.lat",
                                "location.lng",
                                "imageUrl",
                                "creatorId",
                                "createdAt"
                            ],
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "example": [
                            "id",
                            "location"
                        ],
                        "description": "Fields to include in the response; omit for full document",
                        "name": "fields",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "example": [
                            "123456789"
                        ],
                        "description": "The ID of the place",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "example": [
                            "gt:3.5"
                        ],
                        "description": "The latitude of the place",
                        "name": "locationLat",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "example": [
                            "lt:4.5"
                        ],
                        "description": "The longitude of the place",
                        "name": "locationLng",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 1,
                        "description": "The page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 100,
                        "description": "Items per page",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "enum": [
                                "createdAt",
                                "-createdAt",
                                "title",
                                "-title",
                                "description",
                                "-description",
                                "address",
                                "-address"
                            ],
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "example": [
                            "-createdAt"
                        ],
                        "description": "Fields to use for sorting. Use the '-' for descending sorting",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "example": [
                            "eq:Some Place"
                        ],
                        "description": "The place title/name, 10 characters minimum",
                        "name": "title",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.PlacesPaginated"
                        }
                    }
                }
            }
        },
        "/api/places/search": {
            "post": {
                "security": [
                    {
                        "OAuth2Password": [
                            "admin"
                        ]
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Place"
                ],
                "summary": "Search and Retrieve places",
                "parameters": [
                    {
                        "description": "POST parameters",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.PlaceSearch"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.PlacesPaginated"
                        }
                    }
                }
            }
        },
        "/api/places/{placeId}": {
            "get": {
                "security": [
                    {
                        "OAuth2Password": [
                            "admin"
                        ]
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Place"
                ],
                "summary": "Search and Retrieve place by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Place ID",
                        "name": "placeId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.PlaceRead"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "OAuth2Password": [
                            "admin"
                        ]
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Place"
                ],
                "summary": "Update places",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Place ID",
                        "name": "placeId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "PUT parameters",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.PlacePut"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.PlaceRead"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "OAuth2Password": [
                            "admin"
                        ]
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Place"
                ],
                "summary": "Delete place by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Place ID",
                        "name": "placeId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.PlaceDeleteResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "routes.PlaceDeleteResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Deleted place 507f1f77bcf86cd799439011"
                }
            }
        },
...
        "schemas.FlexLocation": {
            "type": "object",
            "properties": {
                "lat": {
                    "description": "The latitude of the place",
                    "type": "number",
                    "example": 51.48180425016331
                },
                "lng": {
                    "description": "The longitude of the place",
                    "type": "number",
                    "example": -0.19090418688755467
                }
            }
        },
        "schemas.Location": {
            "type": "object",
            "properties": {
                "lat": {
                    "description": "The latitude of the place",
                    "type": "number",
                    "example": 51.48180425016331
                },
                "lng": {
                    "description": "The longitude of the place",
                    "type": "number",
                    "example": -0.19090418688755467
                }
            }
        },
        "schemas.PlacePut": {
            "type": "object",
            "properties": {
                "address": {
                    "description": "The place address",
                    "type": "string",
                    "minLength": 1,
                    "example": "Fulham road"
                },
                "description": {
                    "description": "The place description, 10 characters minimum",
                    "type": "string",
                    "minLength": 10,
                    "example": "Stadium of Chelsea football club"
                },
                "location": {
                    "description": "Location object (can be sent as JSON string)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/schemas.FlexLocation"
                        }
                    ]
                },
                "title": {
                    "description": "The place title/name, 10 characters minimum",
                    "type": "string",
                    "minLength": 10,
                    "example": "Stamford Bridge"
                }
            }
        },
        "schemas.PlaceRead": {
            "type": "object",
            "properties": {
                "address": {
                    "description": "The place address",
                    "type": "string",
                    "minLength": 1,
                    "example": "Fulham road"
                },
                "createdAt": {
                    "description": "creation datetime                                            // last update datetime",
                    "type": "string",
                    "example": "2024-01-12T10:15:30.000Z"
                },
                "creatorId": {
                    "description": "The ID of the place creator",
                    "type": "integer",
                    "example": 123456789
                },
                "description": {
                    "description": "The place description, 10 characters minimum",
                    "type": "string",
                    "minLength": 10,
                    "example": "Stadium of Chelsea football club"
                },
                "id": {
                    "description": "The ID of the place",
                    "type": "integer",
                    "example": 123456789
                },
                "imageUrl": {
                    "description": "local url on the storage",
                    "type": "string",
                    "example": "avatar2_80e32f88-c9a5-4fcd-8a56-76b5889440cd.jpg"
                },
                "location": {
                    "description": "Location object (can be sent as JSON string)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/schemas.Location"
                        }
                    ]
                },
                "title": {
                    "description": "The place title/name, 10 characters minimum",
                    "type": "string",
                    "minLength": 10,
                    "example": "Stamford Bridge"
                }
            }
        },
        "schemas.PlaceSearch": {
            "type": "object",
            "properties": {
                "address": {
                    "description": "The place address",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "ilike:Boulevard"
                    ]
                },
                "creatorId": {
                    "description": "The ID of the place creator",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "in:123456789"
                    ]
                },
                "description": {
                    "description": "The place description, 10 characters minimum",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "like:football"
                    ]
                },
                "fields": {
                    "description": "Fields to include in the response; omit for full document",
                    "type": "array",
                    "items": {
                        "type": "string",
                        "enum": [
                            "id",
                            "title",
                            "description",
                            "address",
                            "location.lat",
                            "location.lng",
                            "imageUrl",
                            "creatorId",
                            "createdAt"
                        ]
                    },
                    "example": [
                        "id",
                        "location"
                    ]
                },
                "id": {
                    "description": "The ID of the place",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "123456789"
                    ]
                },
                "locationLat": {
                    "description": "The latitude of the place",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "gt:3.5"
                    ]
                },
                "locationLng": {
                    "description": "The longitude of the place",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "lt:4.5"
                    ]
                },
                "page": {
                    "description": "The page number",
                    "type": "integer",
                    "default": 1,
                    "minimum": 1
                },
                "size": {
                    "description": "Items per page",
                    "type": "integer",
                    "default": 100,
                    "maximum": 100,
                    "minimum": 1
                },
                "sort": {
                    "description": "Fields to use for sorting. Use the '-' for descending sorting",
                    "type": "array",
                    "items": {
                        "type": "string",
                        "enum": [
                            "createdAt",
                            "-createdAt",
                            "title",
                            "-title",
                            "description",
                            "-description",
                            "address",
                            "-address"
                        ]
                    },
                    "example": [
                        "-createdAt"
                    ]
                },
                "title": {
                    "description": "The place title/name, 10 characters minimum",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "eq:Some Place"
                    ]
                }
            }
        },
        "schemas.PlacesPaginated": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schemas.PlaceRead"
                    }
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "totalCount": {
                    "type": "integer",
                    "example": 40
                },
                "totalPages": {
                    "type": "integer",
                    "example": 2
                }
            }
        },
...
    },
    "securityDefinitions": {
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "/api/auth/signin"
        }
    },
...
}